backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
